Prints task name and stack trace on exception
In the AsyncExecutor all of the graph nodes are passed into a
asyncio.wait call. After all tasks are completed the call returns a done
list. This list can then be checked to see if any exceptions happen.
This PR makes that more helpful by printing which op caused the
exception and then also the stack trace from the task.