fix ZerosOp and add rust integration test for ZerosOp (#1046)
This PR adds Rust integration test for ZerosOp and fixes the op.

Specifically, the addition of ZerosOp into pymoose/src/computation.rs was accidentally not committed in the prior PR that added ZerosOp. This PR fixes it and adds a Rust integration test that follows OnesOp Rust integration test implementation.